feat(bundler): support filename template with contenthash
