+++ crates/bundler-core/src/library/ecmascript/mod.rs
@@ -1,6 +1,3 @@
 pub(crate) mod chunk;
-pub(crate) mod content;
-pub(crate) mod version;
 
-pub use chunk::EcmascriptLibraryChunk;
-pub use content::EcmascriptLibraryChunkContent;
+pub use chunk::EcmascriptLibraryEvaluateChunk;
