+++ examples/with-less/project_options.json
@@ -5,7 +5,9 @@
     "entry": [
       {
         "import": "./src/index.ts",
-        "library": {}
+        "library": {
+          "name": "with-less"
+        }
       }
     ],
     "output": {
