+++ examples/with-style-loader/project_options.json
@@ -5,7 +5,9 @@
     "entry": [
       {
         "import": "./src/index.ts",
-        "library": {}
+        "library": {
+          "name": "with-style-loader"
+        }
       }
     ],
     "output": {
